+++ src/main/java/com/nbe8101team03/domain/user/dto/UserInfoRes.java
@@ -3,7 +3,7 @@ package com.nbe8101team03.domain.user.dto;
 import com.nbe8101team03.domain.user.entity.User;
 
 public record UserInfoRes (
-        int user_id,
+        Long user_id,
         String email,
         String address,
         int zipcode
